+++ MercadoPagoSDK/MercadoPagoSDK/Services/Models/PXPayerPaymentMethodDisplayInfo.swift
@@ -6,9 +6,11 @@ open class PXPayerPaymentMethodDisplayInfo: NSObject, Codable {
 
 open class PXDisplayInfoResult: NSObject, Codable {
     let paymentMethod: PXDisplayInfoPaymentMethod?
+    let extraInfo: PXDisplayInfoExtraInfo?
 
     enum CodingKeys: String, CodingKey {
         case paymentMethod = "payment_method"
+        case extraInfo = "extra_info"
     }
 }
 
@@ -18,6 +20,14 @@ open class PXDisplayInfoPaymentMethod: NSObject, Codable {
 
     enum CodingKeys: String, CodingKey {
         case iconUrl = "icon_url"
+        case detail
+    }
+}
+
+open class PXDisplayInfoExtraInfo: NSObject, Codable {
+    let detail: [PXText]?
+
+    enum CodingKeys: String, CodingKey {
         case detail
     }
 }