+++ MercadoPagoSDK/MercadoPagoSDK/UI/PXResult/New UI/Cells/PXNewCustomView.swift
@@ -1,20 +1,23 @@
 import UIKit
+import MLCardDrawer
 
 open class PXNewCustomViewData {
     let firstString: NSAttributedString?
     let secondString: NSAttributedString?
     let thirdString: NSAttributedString?
     let fourthString: NSAttributedString?
+    let bottomString: NSAttributedString?
     let icon: UIImage?
     let iconURL: String?
     let action: PXAction?
     let color: UIColor?
 
-    init(firstString: NSAttributedString?, secondString: NSAttributedString?, thirdString: NSAttributedString?, fourthString: NSAttributedString?, icon: UIImage?, iconURL: String?, action: PXAction?, color: UIColor?) {
+    init(firstString: NSAttributedString?, secondString: NSAttributedString?, thirdString: NSAttributedString?, fourthString: NSAttributedString?, bottomString: NSAttributedString?, icon: UIImage?, iconURL: String?, action: PXAction?, color: UIColor?) {
         self.firstString = firstString
         self.secondString = secondString
         self.thirdString = thirdString
         self.fourthString = fourthString
+        self.bottomString = bottomString
         self.icon = icon
         self.iconURL = iconURL
         self.action = action
@@ -35,7 +38,7 @@ class PXNewCustomView: UIView {
     let data: PXNewCustomViewData
 
     class func getData() -> PXNewCustomViewData {
-        return PXNewCustomViewData(firstString: nil, secondString: nil, thirdString: nil, fourthString: nil, icon: nil, iconURL: nil, action: nil, color: nil)
+        return PXNewCustomViewData(firstString: nil, secondString: nil, thirdString: nil, fourthString: nil, bottomString: nil, icon: nil, iconURL: nil, action: nil, color: nil)
     }
 
     init(data: PXNewCustomViewData, bottomView: UIView? = nil) {
@@ -107,7 +110,7 @@ class PXNewCustomView: UIView {
             }
             if let label = firstLabel {
                 pxContentView.addSubview(label)
-                setTopConstraints(targetView: label, labelsView: labelsView, firstLabel: nil, secondLabel: nil, thirdLabel: nil, actionButton: nil)
+                setTopConstraints(targetView: label, labelsView: labelsView, firstLabel: nil, secondLabel: nil, thirdLabel: nil, fourthLabel: nil, actionButton: nil)
                 NSLayoutConstraint.activate([
                     label.leadingAnchor.constraint(equalTo: labelsView.leadingAnchor),
                     label.trailingAnchor.constraint(equalTo: labelsView.trailingAnchor)
@@ -124,7 +127,7 @@ class PXNewCustomView: UIView {
                     label.leadingAnchor.constraint(equalTo: labelsView.leadingAnchor),
                     label.trailingAnchor.constraint(equalTo: labelsView.trailingAnchor)
                 ])
-                setTopConstraints(targetView: label, labelsView: labelsView, firstLabel: firstLabel, secondLabel: nil, thirdLabel: nil, actionButton: nil)
+                setTopConstraints(targetView: label, labelsView: labelsView, firstLabel: firstLabel, secondLabel: nil, thirdLabel: nil, fourthLabel: nil, actionButton: nil)
             }
         }
 
@@ -138,7 +141,7 @@ class PXNewCustomView: UIView {
                     label.leadingAnchor.constraint(equalTo: labelsView.leadingAnchor),
                     label.trailingAnchor.constraint(equalTo: labelsView.trailingAnchor)
                 ])
-                setTopConstraints(targetView: label, labelsView: labelsView, firstLabel: firstLabel, secondLabel: secondLabel, thirdLabel: nil, actionButton: nil)
+                setTopConstraints(targetView: label, labelsView: labelsView, firstLabel: firstLabel, secondLabel: secondLabel, thirdLabel: nil, fourthLabel: nil, actionButton: nil)
             }
         }
 
@@ -153,7 +156,7 @@ class PXNewCustomView: UIView {
                     fourthLabel.trailingAnchor.constraint(equalTo: labelsView.trailingAnchor)
                 ])
 
-                setTopConstraints(targetView: fourthLabel, labelsView: labelsView, firstLabel: firstLabel, secondLabel: secondLabel, thirdLabel: thirdLabel, actionButton: nil)
+                setTopConstraints(targetView: fourthLabel, labelsView: labelsView, firstLabel: firstLabel, secondLabel: secondLabel, thirdLabel: thirdLabel, fourthLabel: nil, actionButton: nil)
             }
         }
 
@@ -167,11 +170,30 @@ class PXNewCustomView: UIView {
                     button.trailingAnchor.constraint(equalTo: labelsView.trailingAnchor),
                     button.heightAnchor.constraint(equalToConstant: 20)
                 ])
-                setTopConstraints(targetView: button, labelsView: labelsView, firstLabel: firstLabel, secondLabel: secondLabel, thirdLabel: thirdLabel, actionButton: nil)
+                setTopConstraints(targetView: button, labelsView: labelsView, firstLabel: firstLabel, secondLabel: secondLabel, thirdLabel: thirdLabel, fourthLabel: fourthLabel, actionButton: nil)
             }
         }
 
-        if let expectationView = bottomView {
+        var bottomLabel: UILabel?
+        if let bottomString = data.bottomString {
+            bottomLabel = buildBottomLabel(bottomString)
+            if let bottomLabel = bottomLabel {
+                pxContentView.addSubview(bottomLabel)
+                NSLayoutConstraint.activate([
+                    bottomLabel.leadingAnchor.constraint(equalTo: pxContentView.leadingAnchor, constant: PXLayout.L_MARGIN),
+                    bottomLabel.trailingAnchor.constraint(equalTo: pxContentView.trailingAnchor, constant: -PXLayout.L_MARGIN)
+                ])
+
+                if let iconView = iconView, thirdLabel == nil {
+                    bottomLabel.topAnchor.constraint(equalTo: iconView.bottomAnchor, constant: PXLayout.S_MARGIN).isActive = true
+                } else {
+                    setTopConstraints(targetView: bottomLabel, labelsView: labelsView, firstLabel: firstLabel, secondLabel: secondLabel, thirdLabel: thirdLabel, fourthLabel: fourthLabel, actionButton: actionButton)
+                }
+            }
+        }
+
+        if let expectationView = bottomView,
+               bottomLabel == nil {
             pxContentView.addSubview(expectationView)
             NSLayoutConstraint.activate([
                 expectationView.leadingAnchor.constraint(equalTo: pxContentView.leadingAnchor, constant: PXLayout.L_MARGIN),
@@ -181,7 +203,7 @@ class PXNewCustomView: UIView {
             if let iconView = iconView {
                 expectationView.topAnchor.constraint(equalTo: iconView.bottomAnchor, constant: PXLayout.S_MARGIN).isActive = true
             } else {
-                setTopConstraints(targetView: expectationView, labelsView: labelsView, firstLabel: firstLabel, secondLabel: secondLabel, thirdLabel: thirdLabel, actionButton: actionButton)
+                setTopConstraints(targetView: expectationView, labelsView: labelsView, firstLabel: firstLabel, secondLabel: secondLabel, thirdLabel: thirdLabel, fourthLabel: fourthLabel, actionButton: actionButton)
             }
         }
         PXLayout.pinLastSubviewToBottom(view: pxContentView, withMargin: PXLayout.S_MARGIN)
@@ -206,6 +228,15 @@ private extension PXNewCustomView {
         return label
     }
 
+    func buildBottomLabel(_ string: NSAttributedString) -> UILabel {
+        let label = UILabel()
+        label.translatesAutoresizingMaskIntoConstraints = false
+        label.attributedText = string
+        label.lineBreakMode = .byWordWrapping
+        label.numberOfLines = 0
+        return label
+    }
+
     func buildButton(_ action: PXAction) -> UIButton {
         let button = UIButton()
         button.translatesAutoresizingMaskIntoConstraints = false
@@ -217,10 +248,12 @@ private extension PXNewCustomView {
         return button
     }
 
-    func setTopConstraints(targetView: UIView, labelsView: UIView, firstLabel: UILabel? = nil, secondLabel: UILabel? = nil, thirdLabel: UILabel?, actionButton: UIButton? = nil) {
+    func setTopConstraints(targetView: UIView, labelsView: UIView, firstLabel: UILabel? = nil, secondLabel: UILabel? = nil, thirdLabel: UILabel? = nil, fourthLabel: UILabel? = nil, actionButton: UIButton? = nil) {
         var topConstraint: NSLayoutConstraint
         if let actionButton = actionButton {
             topConstraint = targetView.topAnchor.constraint(equalTo: actionButton.topAnchor, constant: PXLayout.XXXS_MARGIN)
+        } else if let fourthLabel = fourthLabel {
+            topConstraint = targetView.topAnchor.constraint(equalTo: fourthLabel.bottomAnchor, constant: PXLayout.XXXS_MARGIN)
         } else if let thirdLabel = thirdLabel {
             topConstraint = targetView.topAnchor.constraint(equalTo: thirdLabel.bottomAnchor, constant: PXLayout.XXXS_MARGIN)
         } else if let secondLabel = secondLabel {