+++ MercadoPagoSDK/MercadoPagoSDK/UI/PXResult/PXPaymentCongratsViewModel.swift
@@ -20,12 +20,18 @@ class PXPaymentCongratsViewModel {
                         thirdString: NSAttributedString?,
                         fourthString: NSAttributedString?)
 
+        var bottomText: NSAttributedString?
         var iconURL: String?
 
         if let displayInfo = paymentInfo.displayInfo {
             subtitles = getSubtitles(from: displayInfo.result?.paymentMethod?.detail)
 
             iconURL = displayInfo.result?.paymentMethod?.iconUrl
+        }
+
+        if let extraInfo = paymentInfo.displayInfo?.result?.extraInfo,
+           let detail = extraInfo.detail {
+            bottomText = getBottomText(detail)
         } else {
             subtitles.secondString = PXNewResultUtil.formatPaymentMethodSecondString(paymentMethodName: paymentInfo.paymentMethodName,
                                                                            paymentMethodLastFourDigits: paymentInfo.paymentMethodLastFourDigits,
@@ -40,6 +46,7 @@ class PXPaymentCongratsViewModel {
                                    secondString: subtitles.secondString,
                                    thirdString: subtitles.thirdString,
                                    fourthString: subtitles.fourthString,
+                                   bottomString: bottomText,
                                    icon: defaultIcon,
                                    iconURL: iconURL,
                                    action: nil,
@@ -58,6 +65,22 @@ class PXPaymentCongratsViewModel {
         return (secondString: secondaryTexts[0],
                 thirdString: secondaryTexts[1],
                 fourthString: secondaryTexts[2])
+    }
+
+    private func getBottomText(_ texts: [PXText]? ) -> NSMutableAttributedString? {
+        var combinedText = NSMutableAttributedString()
+
+        if let texts = texts {
+            for eachText in texts {
+                if let text = eachText.getAttributedString(fontSize: PXLayout.XXS_FONT) {
+                    combinedText.append(text)
+                    if text != texts.last {
+                        combinedText.append(NSAttributedString(string: "\n\n"))
+                    }
+                }
+            }
+        }
+        return combinedText
     }
 }