+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - New payment method model support
+- Added Congrats Extra Info (IFPE regulations)
 
 ### Changed
 - Hide duplicate value when payment method is debit card and splitConfiguration is null